Refactored to move function out of UI components
